plotters: Move `install_plotter` into `plotters`
