+++ .isort.cfg
@@ -43,7 +43,6 @@ extend_skip=
     chia/introducer/introducer.py
     chia/plotters/bladebit.py
     chia/plotters/chiapos.py
-    chia/plotters/install_plotter.py
     chia/plotters/madmax.py
     chia/plotters/plotters.py
     chia/pools/pool_puzzles.py
